backend/requirements.txt
Django>=3.2,<4.0
djangorestframework>=3.12.0,<3.13.0
gunicorn>=20.0.4,<20.1.0
psycopg2>=2.8,<3.0

